requirements.txt
amqp==2.1.4
appdirs==1.4.0
autopep8==1.2.4
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
celery==4.0.2
certifi==2016.9.26
click==6.7
configparser==3.5.0
coreapi==2.1.1
coreapi-cli==1.0.6
Django==1.8
django-angular==0.8.4
django-braces==1.10.0
django-celery-beat==1.0.1
django-highcharts==0.1.7
djangorestframework==3.5.3
docformatter==0.8
enum34==1.1.6
flake8==3.2.1
flower==0.9.1
futures==3.0.5
itypes==1.1.0
kombu==4.0.2
mccabe==0.5.3
packaging==16.8
pep8==1.7.0
pycodestyle==2.2.0
pyflakes==1.3.0
pyformat==0.7
pyparsing==2.1.10
pytz==2016.10
redis==2.10.5
requests==2.12.4
six==1.10.0
tornado==4.2
unify==0.2
untokenize==0.1.1
uritemplate==3.0.0
vine==1.1.3
